Changed zoomLevel to a float

+++ Processing/Main/Screen.pde
@@ -2,7 +2,7 @@ private boolean functionBoardOpen = false;
 private boolean equationsBarOpen = true;
 private int numEquations = 1;
 private int whichEquationSelected = 1;
-private int zoomLevel = 2; // this is the value at the first marked tick on the grid
+private float zoomLevel = 2; // this is the value at the first marked tick on the grid
 
 void topBar() {
   // CREATE DESMOS LOGO ON TOP
